code/backend/requirements.txt
alembic==1.7.5
amqp==5.0.6
aniso8601==9.0.1
attrs==21.2.0
backports.entry-points-selectable==1.1.1
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.0
distlib==0.3.3
filelock==3.4.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
identify==2.4.0
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.2.1
kombu==5.2.2
Mako==1.1.6
MarkupSafe==2.0.1
mypy-extensions==0.4.3
nodeenv==1.6.0
pathspec==0.9.0
platformdirs==2.4.0
pre-commit==2.15.0
prompt-toolkit==3.0.23
pycparser==2.21
PyJWT==2.3.0
PyMySQL==1.0.2
pyrsistent==0.18.0
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.27
toml==0.10.2
tomli==1.2.2
typing-extensions==4.0.0
urllib3==1.26.7
vine==5.0.0
virtualenv==20.10.0
wcwidth==0.2.5
Werkzeug==2.0.2
zipp==3.6.0
